+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * (improvement) Detect space id / content token mismatch and automatically abort all further requests.
 * (feature) Add `storyblok:debug` command.
+* (deprecation) Deprecate `storyblok:components:overview` command, use `storyblok:debug` instead.
 
 
 3.12.1